Merge pull request #2122 from RosettaCommons/lqtza/snugdock_foldtree_update
A frequent complaint of SnugDock users is the requirement for PDB chain order to be LH_A. I'm not sure why this was hard-coded at a point, but it shouldn't be the case. This pull request will update the SnugDock code to use a single FoldTree for docking and loop remodeling. This "universal" FoldTree is shown below.

![snugdock_foldtree](https://user-images.githubusercontent.com/6910140/41170084-f2148f38-6b18-11e8-8808-fc7cc4d8ddcc.png)

It will be set once by setup_ab_ag_foldtree in the apply() function of SnugDockProtocol. This new function will also set up the FoldTree with the appropriate Loops for the CDRs. The universal FoldTree will also set the MoveMap and PackerTask for docking. All these objects will be passed along to the various movers involved in SnugDock, so no additional modification is made, whereas previously the FoldTree would be altered at each step depending.

Future work will involve generalizing the new functions and moving them to AntibodyInfo as there might be some other code that will benefit from this FoldTree.

Old repository SHA1: 2c4ae1785ec09e312edfaefa28b75fe1fa0dd82c

Co-authored-by: Jeliazko Jeliazkov <jeliazkov@jhu.edu>
Co-authored-by: Romero Brissett <rbrissett@buddy.local>