Merge pull request #3190 from RosettaCommons/vmullig/fix_peptidestubmover
The PeptideStubMover is kinda crappy.  Fix it.

This pull request adds proper unit tests for the PeptideStubMover, and fixes the currently super-twitchy behaviour of the mover.  (The problem was diagnosed by Andrew Leaver-Fay, and fixed by Brian Wetizner, so thanks to both of those fellows!)

@weitzner @aleaverfay

Tasks:
- [x] Add unit test for Append.
     - [x] With repeat.
- [x] Add unit test for Prepend.
     - [x] With repeat.
- [x] Fix aberrant behaviour in all of the above.

Not in this pull request, but possibly in future:
- Add unit test for Insert.
     - With repeat.
- Add unit test for sidechain conjugation.

Old repository SHA1: eace34955a8893a9a732f5d67d72930575875216

Co-authored-by: Brian D. Weitzner <brian.weitzner@gmail.com>