Merge pull request #3274 from RosettaCommons/jadolfbr/simple_metrics_custom_type_
Add additional custom_type setting to SimpleMetrics

This adds a `custom_type` option to the `SimpleMetric` framework, which can be set in code or in RosettaScripts.  The `custom_type` allows one to define multiple SimpleMetrics, each with different settings, and have them all be output through a single `RunSimpleMetric` run or reused in multiple.  

For example, you can calculate rmsd for three different atom sets and use `custom_type` to tag them as `BBONLY` `HEAVY`, and `ALL` for example, and output them at the same time using a single mover.  Before this PR, you would need three different RunSimpleMetric movers to change the prefix/suffix in the output.  This reduces the number of times we need to define SMs in RosettaScripts.  

The output name in the scorefile or features db will now be `prefix + custom_type + "_"+ type + suffix`  where the type is the name output by the Metric (such as RMSD, SASA, etc.). If no custom_type is given, we use the same output as before: `prefix + type + suffix`

Old repository SHA1: a3ecb20a8222352a113fc0250b00a861e6145741