Merge pull request #3165 from RosettaCommons/raemisch/DensityScoreResidueSelector
DensityFitResidueSelector

This ResidueSelector  aims to select residues that have good fit (default) /bad fit (invert) into electron density.  Uses either a Zscore made up of density, nbr_density, rama/sugar_bb (for glycans), and cartesian geometry OR the correlation to density that is used internally to determine the density score. 

The Zscore has been adapted and refactored from `FragmentBiasAssigner`, which is used to determine poor density in hybridize.  Thanks to @fdimaio for his help in this!

Numbers and cutoffs match well with Coot's density fit analysis tool.  

Additional options are present to give a residue selector as a mask, which either acts as a subset for `match_res` OR (by default) acts to only calculate the Zscore among those residues (which CAN be turned off).  This is VERY important for glycans or other poorly-resolved residues, or the zscore will be inherently bad for all of them due to the protein bias.

Old repository SHA1: 6f59745c598e398fae372463f1334c8770969d7c

Co-authored-by: raemisch <raemisch@scripps.edu>