Merge pull request #3292 from RosettaCommons/JWLabonte/sugars/database
Carbohydrates: Database additions and .pdb input (and output) bug fixes

This merge attempts to increase the percentage of sugar-containing .pdb files that can be read in to Rosetta without modification.

I have a current benchmark of 104 .pdb files containing (free) glycan ligands. Before these changes, 78/104 files could load. Files that do not load have 5 separate problems, 3 of which I address with this merge. After this merge 92/104 files should load.

As an extra bonus, this merge expands on @BrandonFrenz's work and completes the database such that all sugars can be output with the PDB 3-letter codes, if that is what is desired, not only the limited subset. It also integrates (effectively) the bug fix made in PR #3289 and negates the need for PR #3175.

All unit tests pass.
c-term_conjugation test is simply the centroid ResidueTypeSet reporting on its larger size.
carbohydrates integration test expanded.
glycan_anomers test is currently unstable and it is hoped that this merge corrects that.

Note that master was already broken when I made this merge.

Old repository SHA1: 23b3051a1a53e1fa568ff972ef90a3f86e8caaec