Merge pull request #3260 from RosettaCommons/JWLabonte/sugars/sampling
Carbohydrates: Adding RingPlaneFlipMover

This merge will introduce a new pyranose-specific Mover that will flip an appropriate ring 180 degrees about its anomeric bond with a counter move (similar to how shear moves work) at an equatorial bond at the opposite position on the ring. This move can thus only occur for 4-linked aldopyranose or 5-linked ketopyranose residues, but most natural sugars are aldopyranoses, and the 4-linkage is common. It is expected that this Mover will help sample in cases where a oligo- or poly-saccharide is "stuck" with its ring plane in the wrong orientation, with substituents trapped on the wrong side of the ring.

All tests pass.

Old repository SHA1: c869d56abb713ec2f8552048670acca3247a5423