Merge pull request #3297 from CyrusBiotechnology/indigo/argobody
Code from @indigogo at Cyrus: Antibody grafting, option to define CDRs with an input JSON in exact same format as the preexisting output JSON
(from Indigo:)
Additions to antibody grafting public app. This adds a non-default option to define CDR loop begin/end positions with an input JSON configuration file. If this file is defined on cmdline input, then the RegEx-based CDR prediction is skipped and the CDR begin/end points are defined by the information in the input JSON. This PR adds the following:

a new class to set the CDR sequence position information in AntibodySequence.sequence.cdr.* based on input json numbers. the input json is parsed using the json parsing tools in utility/json_utitlties.hh. JSON format is checked every time and exit with informative message is called if input JSON is misformatted.
control logic in apps/public/antibody.cc to check if -json flag is defined on input, and if it is, skip the RegEx step and do the JSON reading step instead. I am not touching anything else in the antibody.cc code.
The point of all this is to add functionality for connecting any CDR detection tool's output with antibody grafting protocol input.

Old repository SHA1: 51cfd44147445970bfc22489ecdb98924ef8edca

Co-authored-by: Indigo King <indigo@cyrusbio.com>