Merge pull request #3315 from RosettaCommons/hdanderson/protac_stuff
Hdanderson/protac stuff
Converted UBQ_Gp_LYX-Cterm application into a mover compatible with RosettaScripts. Made some modifications to the mover code so that it supported a ResidueSelector for choosing the lysine to which the ubiquitin would bind. Also updated the application so that it calls UBQ_GTPaseMover from protocols.

The output of UBQ_GP_LYX-Cterm is expected to change as part of this merge. There were also some changes in the integration test for cycpep, but I expect that this is just noise. I am open to reverting if the cycpep change turns out to be real.

Old repository SHA1: e8f0fd1d3dfb5cc5e6ae8aab9a670f039d3d7ece