Merge pull request #3403 from CyrusBiotechnology/indigo/argobody_fix_outlier_bug
Antibody grafting, database blast outlier exclusion list, add H3 and orientation

This add 2 lines of code to protocols/antibody/grafting/scs_functor.cc
For the antibody grafting app, blast search of pdb templates can be modified by giving an input list of "oulier" pdbIDs to ecvlude for the search of templates to graft from. The ability to exclude outliers for H3 and orientation was left out of the code (not sure if this was an accident). With this PR, H3 and orientation graft outliers are excluded if and only if you add the corresponding H3 and orientation -specific lines to your outlier list file. This does not change any defaults, but adds the ability to exclude templates for H3 grafting too.

From Cyrus from @indigogo 

Old repository SHA1: 41067c35e93c24d1570cf95766313b712eaeb4a0

Co-authored-by: indigogo <indigo@cyrusbio.com>