Merge pull request #3410 from RosettaCommons/smlewis/PR_patterns
At PRC Day 2, I promised that @weitzner and/or I would write up our standards for PR review. We said we'd put it in the README. We changed our minds, because the README is an obvious target for end users looking for compiling and use documentation, so instead I added it to the https://www.rosettacommons.org/docs/wiki/internal_documentation/GithubWorkflow page and improved the link from here.

I am going to go ahead and merge this on the expectation that most further modifications are to the wiki not the README - but edits here are cheap, so feel free to open another README PR as necessary if there is further comment.

Old repository SHA1: eb949ebdc04feb8fc439ecafa8561d12eb23bcab