Merge pull request #3421 from RosettaCommons/danpf/cyclic_peptide-patch-1
Fix AchiralResidueTypeTestHeaders.hh typo

VKM had a typo in his unit test that resulted in the score of a pose being compared to the score of that same pose (a comparison guaranteed to yield identical results and pass), rather than the intended comparison between the score of the pose and the score of its mirror-image.  This corrects that.

Old repository SHA1: 2a6408cce864954659ff8355f5b36e2f02d76864

Co-authored-by: Daniel Farrell <danpf@uw.edu>