Merge pull request #3500 from RosettaCommons/vmullig/fix_selection_warning
Removing a spurious warning when the last residue in a pose is selected.

Currently, when the last residue of a pose is selected by index, there's a spurious warning printed to the tracer about the selection being outside the range of residues in the pose. This is due to a ">=" comparison that should have been a ">" comparison. This fixes that.

Old repository SHA1: f236a16cef7f693572f06f904ba2caface345ed5