Merge pull request #3526 from RosettaCommons/JackMaguire/MultistageRosettaScriptsSimple
JD3 can now create an archiving directory if needed (very small PR)

JD3 has an option to archive job results on disk instead of in memory. The existing way to do this is to create and empty directory and pass the name of that directory to JD3. I'm adding a hefty 8 lines of code to allow JD3 to create that directory if it does not exist.

Old repository SHA1: a87dceacbaa869bd4f95d2ff4d03b0fb35a14143