Merge pull request #3555 from RosettaCommons/vmullig/fix_a3b
Small fixes to the a3b.crick_params file in the Rosetta database

The refactor of the parametric code added some checks for parameter values, which fail if a Crick params file has r1 values less than 0. I'm correcting the a3b.crick_params file, added by Andy, so that all r1 values are positive. (When flipping the sign of an r1 value, it's necessary to shift omega1 for the same atom by π radians).

This does not change the geometry. This is a minor change that will not affect 99.9% of users.

@everyday847

Old repository SHA1: 712ce0b74308bba9926b88bfbc7046dca2199523