Merge pull request #3559 from RosettaCommons/vmullig/residue_index_description_issue
Resolve an issue with assignment of -1 to an unsigned int.

Tweaking ResidueIndexDescription, which was invalidly assigning -1 to an unsigned int.  Note that this was *not* causing any breakage here, but it's a design pattern that confused the ICC compiler (which threw a warning) and might have confused future developers.  It's probably best to avoid this in order to prevent *human* error.

Old repository SHA1: 2ad300de0f6bc29fee268b2954a07b1222af7c6e