Merge pull request #3562 from RosettaCommons/jadolfbr/sep18
SugarBB Energy fix, Glycan Relax rename

Changes:

1. Add `VIRTUAL_RESIDUE` property to a few patches that were missing them

2. Fix `SugarBBEnergy` to skip virtual residues.  This addresses #3250 

3. Add serialization methods to `NativeAntibodySeq`, simplify its behavior.

4. Rename `GlycanRelaxMover` -> `GlycanTreeSampler`

5. Rename `GlycanTreeRelax`    -> `GlycanTreeModeler`

The renaming has been planned for a while.  Glycan Relax was a place holder for this code.  As we are getting closer to publication, these are the final names. This will break all scripts that use these names in the wild.  Since this is currently unpublished and only the lab and collaborators are using this code, I believe this name change is OK. 

Old repository SHA1: 87ef73394fb6982f2db4e9f968372d436514ddb3