Merge pull request #3658 from RosettaCommons/JackMaguire/MCHBNetSymmFix
Shane Caldwell recently showed me a case where MC HBNet was unable to handle a symmetric pose because of the existence of an additional pure-virtual chain. With this PR, MC HBNet now has bounds-checking to make sure that it does not interact with chains that are not considered subunits.

Thanks to Scott for the review.

Old repository SHA1: dd40f6d03d8da3e96dd1611ac69e21eb569952a8