Merge pull request #3672 from RosettaCommons/jadolfbr/new_simple_metrics
Add 3 new SimpleMetrics with associated unit and integration tests.  Documentation has been updated for this incoming PR. 

ResidueSummaryMetric
=================
The `ResidueSummaryMetric` takes a `PerResidueRealMetric` and summarizes the data in various ways, such as the mean, sum, or the number of residues above, below, or equal to a certain value. This Metric is itself a `RealMetric` and can be used as such in filters, features reporters, etc.  Can use cached metric data. 

InteractionEnergyMetric
=================
A metric for measuring the interaction energy between residues using two sets of residue selectors.  If residue_selector2 is unset, will use TrueResidueSelector to compare 1 with all other residues. `Rama_prepro` and `pro_close` are ignored by default. Options are included for only including a subset of score types or skipping a set of score types. Uses the EnergyMap to find edges and calculates the energy in that way. 

PerResidueClashMetric
=================
A SimpleMetric that calculates the number of atomic clashes per residue using the LJ radius (at 0) using two residue selectors. Default is to use a soft radius, which reduces it by 33% (the dampening of which can be set.).  Does not calculate intra-residue clashes. 


Other
====
Fix serialization of the SimpleMetric custom_type option. 
Add an `is_bonded` function to Conformation and associated unit test.
Change to print out full name of SimpleMetric during RunSimpleMetric's apply function.

Old repository SHA1: e1185d22633f69b3538d32bd2d6ea2535deafe38