Merge pull request #3685 from RosettaCommons/JackMaguire/Ofast
Added "mode=release_native" build option

This might be deemed trivial, but I have measured that building with -Ofast instead of -O3 results in a 2.5% speedup in Rosetta runtime (i.e., the new runtime is 97.5% of the default runtime).

mode=release_native also adds the -mtune="native" flag which allows the compiler to optimize for the specific processor being used to compile. This means that you would not want to build on one computer and run it on a different computer (though you could get away with it if the CPUs are similar enough).

I'm not submitting all tests for this because this change wouldn't affect any of them. I'll add the build test just to make sure that the build process still works though.

Thanks to Sergey for the review!

Old repository SHA1: 712dcfa784d66b2c7d0532d84a578f7497b199e7