Merge pull request #3753 from RosettaCommons/JackMaguire/ExFlagProfileTest2
This corrects behavior from #3745

I added a very short profile test that was intended to track memory usage and not focus on speed. Turns out that it is too short for the benchmark server and now it always fails with the warning "Execution time is below 128s, marking test as failed!". This PR increases the nstruct of that test so that it takes longer (still only a few minutes).

Old repository SHA1: eea2a5b248f13479710959c3cceaedc8d0ca9d1a