Merge pull request #3776 from RosettaCommons/roccomoretti/upper_atom_selector
Convert UPPER_POSITION patch selector to UPPER_ATOM selector.

The UPPER_POSITION patch selector currently has a dependency on the CarbohydrateInfo object. This makes the selector not as general as it could be. I've changed this to an UPPER_ATOM selector, which cues based on the atom name of the atom connected to the UPPER connection, rather than the mainchain position setting in the CarbohydrateInfo object.

Old repository SHA1: ec19188ee3791de9fecd9d148263bba795accb18

Co-authored-by: Labonte <JWLabonte@jhu.edu>