Merge pull request #3779 from RosettaCommons/roccomoretti/change_position_of_atom_on_ring
Fix for position_of_atom_on_ring()

There was some confusion on the behavior of position_of_atom_on_ring() in the case when an atom is part of the ring. The documentation was unclear, and the actual implementation was highly dependent on the atom ordering of get_adjacent_heavy_atoms(). This PR changes things such that atoms on the ring get a return value of 0, just like atoms which aren't directly connected to the ring.

Old repository SHA1: 60d3fe8b4d2d9372d8d8826192a564cca8029700