Merge pull request #3788 from RosettaCommons/JackMaguire/Allocations
There was a needless Residue copy in the inner loop of the FixbbSimAnnealer. It accounted for 54% of all heap allocations and 2% of total runtime (25% of total runtime according to the benchmark server).

Thanks to Vikram for the review!

Old repository SHA1: 95d4d5c4f8c3c09c2ad3ddcfa6f96f2a183c42f8