Merge pull request #3798 from RosettaCommons/jadolfbr/sm_in_xmlobjects
This PR adds functions to access individual `SimpleMetrics` from the `XmlObject` class - which is useful for PyRosetta.  The SM can call the apply function of the base SM, which will add the data to the pose.  With #3797 it should now be easy to access the calculated SM data as a dict. 

It should be noted that some functions, like `cached_calculate` and `calculate`, are not in the base class, and so, cannot be accessed in this way in PyRosetta without knowing which type of SM you are obtaining ahead of time.  In C++, you can cast it.

Unit tests are updated to fully test these additions.

Old repository SHA1: 4194b26a048bfa344cbd83cc1e1ff036256dfd85