Merge pull request #3836 from RosettaCommons/jadolfbr/fix_sm_filter
This fixes the Sum option for the `SimpleMetricFilter` in parse_my_tag, which is used for `PerResidueMetrics`.  It also adds output for action options to the doc XSD doc string in the `ResidueSummaryMetric`, which is a meta-metric that calculates different values for `PerResidueMetrics`.  

Thanks to @tjbrunette for identifying these issues!

Old repository SHA1: ac47f324a4e2081ea40093d27efa4e283c7f1f4e